Reworking the app to use JWT Authentication/Authorization.
There is still a lot of work to be done here, but it should
be mostly functional. This refactor broke a lot of tests and
those all had to be repaired.  Another consequence is that
test coverage is less complete now and I will need to address
that when I can.

CollectionBase guards againsts trying to retrieve collections
when logged out.  This prevents error spew in the console
that helped to obscure other issues, so it seemed prudent
to block it from happening.

I added a new Button component primarily for the Login
modal. I expect to be able to use it in other places.

Another consequence of this is that checking login status
can be a lot simpler for now.  The tokens don't expire at
this point, so the mere existence of the token is sufficient
to assume being logged in.  I need to add expiration and
occasional token refreshing.