Possibly fix some test setup stuff.
I have had a few tests that seemed to be flaky because of test
setup.  In particular, it seemed that mocks were getting into
an inconsistent state.  To this end, I've added a jest.config.json
with flags explicitly set even if the value is the same as the
stated default in the documentation.